Fixed format specifier
git-svn-id: https://smlnj-gforge.cs.uchicago.edu/svn/manticore/branches/swp@4199 ac4fbcad-d21d-0410-b760-c7b1d3380187

+++ src/tools/log-dump/log-dump.cxx
@@ -154,7 +154,8 @@ static void LoadLogFile (LogFileDesc *logFileDesc, const char *file)
 	exit (1);
     }
     if (Hdr->hdrSzB != sizeof(LogFileHeader_t)) {
-	fprintf(stderr, "bogus header size %d (expected %d)\n", Hdr->hdrSzB, sizeof(LogFileHeader_t));
+	fprintf(stderr, "bogus header size %d (expected %d)\n",
+	    Hdr->hdrSzB, (int)sizeof(LogFileHeader_t));
 	exit (1);
     }
     if (Hdr->majorVersion != LOG_VERSION_MAJOR) {
